cpuidle: idle2: v0.200 standalone patch The majority of this code is a port from the 2.6.32 P1000 kernel, much of which was required to be rewritten or substantially altered by myself.
Some inspiration and a small amount of code (clock/power gating) was taken
from deep idle by ezekeel. Some additional inspiration was taken from the
nvidia tegra3 cpuidle.

2 idle states have been fully implemented and are available at all
times. The states are IDLE and IDLE2, which is DEEP-IDLE TOP OFF.
DEEP-IDLE TOP ON is not implemented, as it has no powersaving benefit.

Improvements over deep idle are:
It requires no configuration and is enabled automatically
It does not require the cpu to be at 800MHz for stability

A workqueue is used to inform idle2 of events from other subsystems,
such as audio, usb connected and bluetooth. This avoids the need to
continually poll.
PM notifiers are used to notify idle2 of a suspend, although it should
*never* get into a situation where it is enabled at the same time as the
system is suspending.

The IDLE2 mode will only be activated when audio has been playing back
for more than 30 seconds, the device is not connected to USB, bluetooth is
not active (it can be enabled, but not be actively working) various clocks
are gated (in practice, this means that the screen has to be off), certain
subsystems are not active and there are no interrupts pending.

IDLE2 can be disabled with the following command:
echo 1 > /sys/module/cpuidle/parameters/idle2_disabled

Stats can be obtained from the standard cpuidle interface, which, for
IDLE2 is located at:
/sys/devices/system/cpu/cpu0/cpuidle/state1