omap4: Add basic earlysuspend hotplug driver
This driver hotplugs CPU1 offline when early_suspend is called
and onlines it when late_resume is called.
This is advantageous to power saving as CPU0 can then enter
OSWR without agreeing with CPU1, which saves power in certain
scenarios such as screen off audio playback and also when the
device is woken by a partial wakelock to perform a routine task.